workaround for transpilation issues
When working with modules, we either are in an env of:
- node
- browser

In node, we transpile from: typescript -> es5/commonjs
In browser, we transpile from: typescript -> esnext/esnext -> babel -> es5/commonjs

Unfortunately typescript imports non modules (those that use module.export = {}) differently than babel, using the syntax `import * as moment from 'moment'` vs babel which uses syntax `import moment from 'moment'`.  If we only ran from node or browser, we could pick one, but since this is a universal app we need both env to work.  The workaround is to import both and choose the syntax that loaded correctly.

Ideally we could make node work similar to the browser, or all packages exported modules consistently, but both have proven to be very difficult, so here we are :)